requirements.txt
numpy
pandas
Pillow
ipdb
tensorflow-gpu
opencv-python
requests-futures
